split up files, start css cleanup
