+++ client/scripts/views/components/stage_editor.tsx
@@ -1,234 +1,16 @@
 /* @jsx m */
 
 import m from 'mithril';
-import { uuidv4 } from 'lib/util';
-import {
-  QueryList,
-  ListItem,
-  Button,
-  Classes,
-  Dialog,
-  Icons,
-} from 'construct-ui';
+import { Button, Classes, Dialog, Icons } from 'construct-ui';
 
 import 'components/stage_editor.scss';
 
 import app from 'state';
 import { offchainThreadStageToLabel, parseCustomStages } from 'helpers';
 import { ChainEntity, OffchainThread, OffchainThreadStage } from 'models';
-import { chainEntityTypeToProposalName } from 'identifiers';
-import { EntityRefreshOption } from 'controllers/server/chain_entities';
-import {
-  loadMultipleSpacesData,
-  SnapshotProposal,
-} from 'helpers/snapshot_utils';
-
-type SnapshotProposalSelectorAttrs = {
-  onSelect: () => void;
-  snapshotProposalsToSet: SnapshotProposal[];
-  thread: OffchainThread;
-};
-
-class SnapshotProposalSelector
-  implements m.ClassComponent<SnapshotProposalSelectorAttrs>
-{
-  private all_proposals: SnapshotProposal[];
-  private initialized: boolean;
-  private snapshotProposalsLoaded: boolean;
-
-  view(vnode) {
-    const { onSelect } = vnode.attrs;
-
-    if (!app.chain || !app.activeChainId()) return;
-
-    if (!this.initialized) {
-      this.all_proposals = [];
-      this.initialized = true;
-
-      loadMultipleSpacesData(app.chain.meta.chain.snapshot).then((data) => {
-        for (const { proposals } of data) {
-          this.all_proposals = [...this.all_proposals, ...proposals];
-        }
-
-        this.snapshotProposalsLoaded = true;
-        m.redraw();
-      });
-    }
-
-    return (
-      <div class="ChainEntitiesSelector">
-        {this.snapshotProposalsLoaded ? (
-          <QueryList
-            checkmark={true}
-            items={this.all_proposals.sort((a, b) => {
-              return b.created - a.created;
-            })}
-            inputAttrs={{
-              placeholder: 'Search for an existing snapshot proposal...',
-            }}
-            itemRender={(sn: SnapshotProposal) => {
-              const selected = sn.id === vnode.attrs.thread.snapshotProposal;
-              // TODO: show additional info on the ListItem,
-              // like any set proposal title, the creator, or other metadata
-              return (
-                <ListItem
-                  label={
-                    <div class="chain-entity-info">
-                      <div class="chain-entity-top">
-                        {sn.title.slice(0, 60)}...
-                      </div>
-                      <div class="chain-entity-bottom">Hash: ${sn.id}</div>
-                    </div>
-                  }
-                  selected={selected}
-                  key={sn.id}
-                />
-              );
-            }}
-            itemPredicate={(query, sn: SnapshotProposal) => {
-              // TODO
-              return sn.title
-                ?.toString()
-                .toLowerCase()
-                .includes(query.toLowerCase());
-            }}
-            onSelect={(sn: SnapshotProposal) => {
-              onSelect(sn);
-            }}
-          />
-        ) : (
-          <div class="chain-entities-selector-placeholder">
-            <div class="chain-entities-selector-placeholder-text">
-              {this.snapshotProposalsLoaded
-                ? 'TODO: how to begin?'
-                : 'Loading snapshot proposals...'}
-            </div>
-          </div>
-        )}
-      </div>
-    );
-  }
-}
-
-type ChainEntitiesSelectorAttrs = {
-  chainEntitiesToSet: ChainEntity[];
-  onSelect: () => void;
-  thread: OffchainThread;
-};
-
-class ChainEntitiesSelector
-  implements m.ClassComponent<ChainEntitiesSelectorAttrs>
-{
-  private chainEntitiesLoaded: boolean;
-  private initialized: boolean;
-
-  view(vnode) {
-    const { thread, onSelect } = vnode.attrs;
-
-    if (!app.chain || !app.activeChainId()) return;
-
-    if (!this.initialized) {
-      this.initialized = true;
-      app.chain.chainEntities
-        ?.refresh(app.chain.id, EntityRefreshOption.AllEntities)
-        .then(() => {
-          // refreshing loads the latest chain entities into app.chain.chainEntities store
-          this.chainEntitiesLoaded = true;
-          m.redraw();
-        });
-    }
-
-    return (
-      <div class="ChainEntitiesSelector">
-        {this.chainEntitiesLoaded ? (
-          <QueryList
-            checkmark={true}
-            items={app.chain.chainEntities.store.getAll().sort((a, b) => {
-              if (!a.threadId && b.threadId) return -1;
-              if (a.threadId && !b.threadId) return 1;
-              return 0;
-            })}
-            inputAttrs={{
-              placeholder: 'Search for an existing proposal...',
-            }}
-            itemRender={(ce: ChainEntity) => {
-              const selected =
-                vnode.attrs.chainEntitiesToSet
-                  .map((ce_) => ce_.id)
-                  .indexOf(ce.id) !== -1;
-              // TODO: show additional info on the ListItem,
-              // like any set proposal title, the creator, or other metadata
-              return (
-                <ListItem
-                  disabled={ce.threadId && ce.threadId !== thread.id}
-                  label={
-                    <div class="chain-entity-info">
-                      <div class="chain-entity-top">
-                        {chainEntityTypeToProposalName(ce.type) +
-                          (ce.typeId.startsWith('0x')
-                            ? ` ${ce.typeId.slice(0, 6)}...`
-                            : ` #${ce.typeId}`)}
-                      </div>
-                      <div class="chain-entity-bottom">
-                        {ce.threadTitle !== 'undefined'
-                          ? decodeURIComponent(ce.threadTitle)
-                          : ''}
-                      </div>
-                    </div>
-                  }
-                  selected={selected}
-                  key={ce.id ? ce.id : uuidv4()}
-                />
-              );
-            }}
-            itemPredicate={(query, ce: ChainEntity) => {
-              if (ce.typeId.startsWith('0x')) {
-                return false;
-              } else {
-                return (
-                  ce.typeId
-                    .toString()
-                    .toLowerCase()
-                    .includes(query.toLowerCase()) ||
-                  ce.title
-                    ?.toString()
-                    .toLowerCase()
-                    .includes(query.toLowerCase()) ||
-                  chainEntityTypeToProposalName(ce.type)
-                    .toLowerCase()
-                    .includes(query.toLowerCase())
-                );
-              }
-            }}
-            onSelect={(ce: ChainEntity) => {
-              if (
-                vnode.attrs.chainEntitiesToSet
-                  .map((ce_) => ce_.id)
-                  .indexOf(ce.id) !== -1
-              ) {
-                const index = vnode.attrs.chainEntitiesToSet.findIndex(
-                  (ce_) => ce_.id === ce.id
-                );
-                vnode.attrs.chainEntitiesToSet.splice(index, 1);
-              } else {
-                vnode.attrs.chainEntitiesToSet.push(ce);
-              }
-              onSelect(ce);
-            }}
-          />
-        ) : (
-          <div class="chain-entities-selector-placeholder">
-            <div class="chain-entities-selector-placeholder-text">
-              {this.chainEntitiesLoaded
-                ? 'Select "In Voting" to begin.'
-                : 'Loading on-chain proposals...'}
-            </div>
-          </div>
-        )}
-      </div>
-    );
-  }
-}
+import { SnapshotProposal } from 'helpers/snapshot_utils';
+import { SnapshotProposalSelector } from './snapshot_proposal_selector';
+import { ChainEntitiesSelector } from './chain_entities_selector';
 
 type StageEditorAttrs = {
   onChangeHandler: () => void;
@@ -238,10 +20,10 @@ type StageEditorAttrs = {
 };
 
 export class StageEditor implements m.ClassComponent<StageEditorAttrs> {
-  chainEntitiesToSet: ChainEntity[];
-  isOpen: boolean;
-  snapshotProposalsToSet: SnapshotProposal[];
-  stage: OffchainThreadStage;
+  private chainEntitiesToSet: ChainEntity[];
+  private isOpen: boolean;
+  private snapshotProposalsToSet: SnapshotProposal[];
+  private stage: OffchainThreadStage;
 
   oninit(vnode) {
     this.isOpen = !!vnode.attrs.popoverMenu;
