+++ client/styles/components/stage_editor.scss
@@ -30,55 +30,3 @@
     padding: 4px 15px 3px;
   }
 }
-
-.ChainEntitiesSelector {
-  border-radius: 5px;
-  border: 1px solid #ddd;
-  margin: 10px 8px 20px;
-  padding: 0;
-
-  .cui-control-group {
-    margin-bottom: 0;
-  }
-
-  .cui-input input {
-    border-bottom-color: $border-color-light;
-    border-left: none;
-    border-radius: 0;
-    border-right: none;
-    border-top: none;
-    height: 44px;
-  }
-
-  .cui-list {
-    border-radius: 5px;
-    max-height: 150px;
-    min-height: 150px;
-
-    .cui-active {
-      background: initial;
-    }
-
-    .cui-list-item {
-      .chain-entity-info {
-        .chain-entity-bottom {
-          font-size: 88%;
-        }
-      }
-    }
-  }
-
-  .cui-query-list-checkmark .cui-list-item-content-left {
-    left: 0;
-  }
-
-  .chain-entities-selector-placeholder {
-    padding: 112px 0;
-    text-align: center;
-  }
-
-  .chain-entities-selector-placeholder-text {
-    color: $text-color-medium-dark;
-    height: 30px;
-  }
-}
